Preliminary clean-up in preparation for PR raising.
