+++ .idea/compiler.xml
@@ -13,6 +13,7 @@
       <module name="behave_main" target="1.8" />
       <module name="behave_rpcProxy" target="1.8" />
       <module name="behave_scenario" target="1.8" />
+      <module name="behave_smokeTest" target="1.8" />
       <module name="behave_test" target="1.8" />
       <module name="bootstrapper_main" target="1.8" />
       <module name="bootstrapper_test" target="1.8" />
