+++ experimental/behave/src/main/kotlin/net/corda/behave/node/Distribution.kt
@@ -166,11 +166,6 @@ class Distribution private constructor(
             "master"  -> LATEST_MASTER
             "r3-master"  -> LATEST_R3_MASTER
             "corda-3.0" -> fromArtifactory(Type.CORDA, version)
-            "corda-3.0-HC02" -> fromArtifactory(Type.CORDA, version)
-            "corda-3.0-RC01" -> fromArtifactory(Type.CORDA, version)
-            "corda-3.0-RC02" -> fromArtifactory(Type.CORDA, version)
-            "corda-3.0-RC03" -> fromArtifactory(Type.CORDA, version)
-            "corda-3.0-pre-release-V3" -> fromJarFile(Type.CORDA,"corda-3.0-pre-release-V3")
             "r3corda-3.0" -> R3_V3
 //            "r3corda-3.0-DP2" -> fromArtifactory(DistributionType.R3_CORDA,"R3.CORDA-3.0.0-DEV-PREVIEW-2")
             "r3corda-3.0-DP2" -> fromJarFile(Type.R3_CORDA,"R3.CORDA-3.0.0-DEV-PREVIEW-2")
