+++ experimental/behave/src/main/kotlin/net/corda/behave/node/Node.kt
@@ -85,8 +85,10 @@ class Node(
         log.info("Configuring {} ...", this)
         serviceDependencies.addAll(config.database.type.dependencies(config))
         config.distribution.ensureAvailable()
-//        config.writeToFile(rootDirectory / "${config.name}.conf")
-        config.writeToFile(rootDirectory / "${config.name}"/ "node.conf")
+        if (config.distribution.type == Distribution.Type.CORDA)
+            config.writeToFile(rootDirectory / "${config.name}_node.conf")
+        else
+            config.writeToFile(rootDirectory / "${config.name}"/ "node.conf")
         installApps()
     }
