+++ experimental/behave/src/qa-scenarios/resources/features/compatibility/network-management/nodeinfo-change.feature
