+++ experimental/behave/src/smoke-test/kotlin/net/corda/behave/service/proxy/CordaRPCProxyClientTest.kt
@@ -1,7 +1,6 @@
 package net.corda.behave.service.proxy
 
 import net.corda.behave.network.Network
-import net.corda.behave.node.Distribution
 import net.corda.behave.node.configuration.NotaryType
 import net.corda.core.messaging.startFlow
 import net.corda.core.utilities.NetworkHostAndPort
